Bring upto speed with Ghost 1.5.0.
As an experiment, was trying to integrate Angular with Ghost and used this repo as a starting point. Cleaned things up per feedback from GScan (https://gscan.ghost.org/). However, it looks like the original code never defined the loops partial template that author.hbs and tag.hbs are looking for. So, this code is not functional even after these changes.